+++ SciProfile/latex.ps1
@@ -16,12 +16,17 @@ function Global:Move-LaTeXBuild {
         [System.String] $Path,
 
         [Parameter(Position=2, Mandatory=$True)]
-        [System.String] $Build
+        [System.String] $Build,
+
+        [Parameter(Position=3, Mandatory=$True)]
+        [System.String] $Destination,
+
+        [Switch] $Copy
     )
 
     Process {
 
-        $libraryLfs = Join-Path -Path "A:\Dropbox\Literature\Documents-Pdf" -ChildPath (Split-Path -Path $Path -Leaf)
+        $libraryLfs = Join-Path -Path $Destination -ChildPath (Split-Path -Path $Path -Leaf)
         $buildDirectory = $Build
 
         If (-not (Test-Path $buildDirectory) -and -not (Test-Path $libraryLfs)) {
@@ -42,18 +47,19 @@ function Global:Move-LaTeXBuild {
             Copy-Item -Path $_ -Destination $destination -Force
         }
 
+        if (-not $Copy){
+            Get-ChildItem -Path $buildDirectory -Directory | ForEach-Object{
 
-        Get-ChildItem -Path $buildDirectory -Directory | ForEach-Object{
+                $destination = Join-Path -Path $libraryLfs -ChildPath (Split-Path -Path $_ -Leaf)
+                If (Test-Path $destination) {
+                    Remove-Item -Path $destination -Recurse -Force
+                }
 
-            $destination = Join-Path -Path $libraryLfs -ChildPath (Split-Path -Path $_ -Leaf)
-            If (Test-Path $destination) {
-                Remove-Item -Path $destination -Recurse -Force
+                Write-FormattedProcess -Message "[MV] '$_' to '$destination"
+
+                
+                Move-Item -Path $_ -Destination $destination -Force
             }
-
-            Write-FormattedProcess -Message "[MV] '$_' to '$destination"
-
-            Move-Item -Path $_ -Destination $destination -Force 
         }
-
     }
 }
